+++ src/lib/manager/AutoHandler.ts
@@ -26,13 +26,18 @@ export async function handleAutoCommand(client: Client, msg: Message): Promise<B
 
     if (msg.channel.type == "text") {
         if (!((msg.channel as TextChannel).permissionsFor(client.user) as Permissions).has(autoResponse.permissions as PermissionResolvable)) return false;
+
+        let serverManager = client.serverManager;
+        let server = await serverManager.getServer(msg.guild.id);
+        let commandRestrictions = server.getCommandRestrictions(msg.channel.id);
+
+        if ((commandRestrictions as Array<string>).includes(autoResponse.name)) return false;
     }
 
     try {
         await autoResponse.run(client, msg);
         console.log(`${autoResponse.name} was initiated sent by ${msg.author.tag}/${msg.author.id}.`);
         client.emit("log", (`User ${msg.author.id}/${msg.author.tag} initiaited autoresponse ${autoResponse.name} in ${(msg.guild) ? `guild ${msg.guild.id}/${msg.guild.name}` : "dms"}.`));
-        msg.channel.stopTyping(true);
     } catch (err) {
         console.warn("Error while executing autoresponse " + autoResponse.name, err);
         client.emit("error", err, autoResponse.name);