+++ src/preload/commands/language/delete/index.ts
@@ -3,7 +3,7 @@ import { Client } from "../../../../lib/types";
 import { Message, MessageEmbed } from "discord.js";
 import colors from "../../../../lib/misc/colors";
 
-export default class Delete extends Command {
+export default class LanguageDelete extends Command {
     constructor() {
         super(
             "delete",
@@ -19,20 +19,19 @@ export default class Delete extends Command {
     }
 
     async run(client: Client, msg: Message): Promise<boolean> {
+        if(!msg.member.permissions.has(["ADMINISTRATOR"])) return false;
         let embed = new MessageEmbed();
-        const lang = await this.getLanguage(msg);
 
         const server = await client.serverManager.getServer(msg.guild.id);
-        
+        server.server.language = 'en-US';
+        await server.save();
+
+        const lang = await this.getLanguage(msg);
         embed.setAuthor(msg.author.tag, msg.author.displayAvatarURL());
         embed.setColor(colors.red);
         embed.setDescription(lang.language_delete_success);
         embed.setTitle(lang.language_delete_embedfooter.replace('%author', msg.author.tag));
         embed.setTimestamp();
-
-        server.server.language = 'en-US';
-
-        await server.save();
 
         await msg.channel.send(embed);
         return true;